refactor: 스케줄러 실행 메서드 클래스로 분리
