+++ src/main/java/site/dogether/challengegroup/service/scheduler/ChallengeGroupStatusScheduler.java
@@ -3,16 +3,15 @@ package site.dogether.challengegroup.service.scheduler;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
-import site.dogether.challengegroup.service.ChallengeGroupService;
 
 @RequiredArgsConstructor
 @Component
 public class ChallengeGroupStatusScheduler {
 
-    private final ChallengeGroupService challengeGroupService;
+    private final ChallengeGroupStatusUpdateService challengeGroupStatusUpdateService;
 
     @Scheduled(cron = "0 0 0 * * *")
     public void runStatusUpdate() {
-        challengeGroupService.updateChallengeGroupStatus();
+        challengeGroupStatusUpdateService.updateChallengeGroupStatus();
     }
 }
